refactor: Modal UI 역할이 사라져 StartButtonContainer 제거 #272
