+++ frontend/src/components/StartButton/StartButton.test.tsx
@@ -2,14 +2,14 @@ import { screen, waitFor } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { http, HttpResponse } from 'msw';
 
-import StartButtonContainer from '../StartButtonContainer';
+import StartButton from './StartButton';
 
 import { ERROR_MESSAGE } from '@/constants/message';
 import { MOCK_API_URL } from '@/constants/url';
 import { server } from '@/mocks/server';
 import { customRenderWithIsMaster } from '@/utils/test-utils';
 
-describe('StartButtonContainer', () => {
+describe('StartButton 테스트', () => {
   it('시작 버튼을 클릭했을 때, 게임 시작 API에서 에러가 발생하면 알림 모달이 뜬다.', async () => {
     const user = userEvent.setup();
     server.use(
@@ -24,7 +24,7 @@ describe('StartButtonContainer', () => {
       }),
     );
 
-    customRenderWithIsMaster(<StartButtonContainer />, true);
+    customRenderWithIsMaster(<StartButton />, true);
 
     const startButton = await screen.findByRole('button', { name: '시작' });
     await user.click(startButton);
