+++ frontend/src/components/StartButton/hooks/useGameStart.ts
