--- frontend/src/components/StartButtonContainer/StartButtonContainer.tsx
@@ -1,7 +1,0 @@
-import StartButton from './StartButton/StartButton';
-
-const StartButtonContainer = () => {
-  return <StartButton />;
-};
-
-export default StartButtonContainer;
